Created a Deployment Version and began scaling and prefab construction
The Hopewell Deployment folder is a stripped down version of the main
Hopewell folder and is thus a smaller version.

The Barrack and Bathroom have been properly scaled to the 1 meter = 1
unit scale in Unity and a third person player has been added to the
scene for adjustment.

Creation of Prefabs from these scaled Barracks and Bathrooms is now
underway with Barracks (2 Story) completed.

Model Barrack and Bathroom that have been scaled are static objects to
prevent changes.